requirements.txt
streamlit
geopandas
rasterio
pyproj
osmnx
shapely
scipy
scikit-image
matplotlib
pandas
numpy
earthengine-api
geemap
leafmap
localtileserver
xarray
plotly
fpdf2
folium
tqdm
streamlit-folium
contextily